requirements.txt
fastapi
uvicorn[standard]
pydantic
yfinance
pandas
numpy
scikit-learn
tensorflow
keras
requests
transformers
torch
PyMuPDF
pdf2image
pytesseract
langchain
langchain-google-genai
google-generativeai
python-dotenv
motor
passlib[bcrypt]